pyproject.toml
[tool.poetry]
name = "nonebot-plugin-flexperm"
version = "0.1.0"
description = ""
authors = ["Muchan <liuzh1773@buaa.edu.cn>"]

[tool.poetry.dependencies]
python = "^3.7.3"
"ruamel.yaml" = "^0.17.10"

[tool.poetry.dev-dependencies]
nonebot2 = "^2.0.0a13"
nonebot-adapter-cqhttp = "^2.0.0a13"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
